[클린코드 7기 이승] 점심 뭐 먹지 미션 STEP 3 (#16)
* feat: 즐겨 찾기 아이콘 추가가

* feat: 모달에서 즐겨찾기 추가

* feat: 즐겨찾기 아이콘 카드에서도 동시에 변경되도록 수정

* fix: data 안나오는 오류 수정

* fix: update 에러 수정

+++ src/model/saveData.js
@@ -1,8 +1,16 @@
 import data from '../data.js'
 
 export const getInitData = () => {
-	localStorage.setItem('restaurantData', JSON.stringify(data.storeList))
-	return JSON.parse(localStorage.getItem('restaurantData'))
+	try {
+		const existingData = localStorage.getItem('restaurantData')
+		if (!existingData) {
+			localStorage.setItem('restaurantData', JSON.stringify(data.storeList))
+		}
+		return JSON.parse(localStorage.getItem('restaurantData'))
+	} catch (error) {
+		console.error('LocalStorage 접근 오류:', error)
+		return data.storeList
+	}
 }
 
 export const addData = (newItem) => {
